Happy Flow Done. Completed All the Navigations Handled Cart Basic and Many more

+++ app/src/main/java/com/intsab/intsabshop/data/Models/ProductItem.java
@@ -40,7 +40,7 @@ public class ProductItem {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(double price) {
         this.price = price;
     }
